requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
distro==1.9.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.11.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
narwhals==2.8.0
numpy==2.3.4
openai==2.4.0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
protobuf==6.33.0
pyarrow==21.0.0
pydantic==2.12.2
pydantic_core==2.41.4
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
referencing==0.37.0
regex==2025.9.18
requests==2.32.5
rpds-py==0.27.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.50.0
tenacity==9.1.2
tiktoken==0.12.0
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
